fix: Added `isCustomElement` functionality to `MDCRenderer`.
Fixed an issue where custom tags would generate a warning despite
defining an exeption in vue.compilerOptions.isCustomElement.

In order to use custom elements in MDC markdown, the user must define a
function in `utils/compilerOptions/isCustomElement.ts` as the default
function. This is then read my `MDCRenderer` which in turn will not
generate the warning.